Faltó algo del anterior commit. Listo.

+++ TP4/src/ejercicio3/FrameEjercicio3.java
@@ -168,7 +168,7 @@ public class FrameEjercicio3 extends JFrame {
 		String input = jtf.getText();
 		if (input.length() > 0) {
 			try {
-				float horas = Float.valueOf(input).intValue();
+				float horas = Float.valueOf(input).floatValue();
 				if (horas < 0 || horas > 24) {
 					JOptionPane.showMessageDialog(null, "La cantidad de horas ingresada esta fuera del rango permitido.\n Debe ser mayor o igual a 0 y menor o igual que 24");
 					System.out.println("Cant. de horas no valida");
